pyproject.toml
[project]
name = "jnkn"
version = "0.1.0"
description = "CI/CD Gatekeeper for Infrastructure and Data Impact Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "networkx>=3.2.1",
    "openlineage-python>=1.40.1",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
full = [
    "sqlglot>=20.0.0",
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.2",
    "httpx>=0.27.0",
    "xxhash>=3.4.1",
]
docs = [
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
    "mkdocs-redirects",
    "mike",
]

[project.scripts]
jnkn = "jnkn.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jnkn"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.0",
    "mkdocstrings-python>=2.0.1",
    "pymdown-extensions>=10.18",
]
